rework the new pool into TransactionPool and Pool impls
