+++ pool/src/lib.rs
@@ -22,8 +22,9 @@
 #![warn(missing_docs)]
 
 pub mod types;
+mod pool;
 pub mod blockchain;
-pub mod minimal_pool;
+pub mod transaction_pool;
 
 extern crate blake2_rfc as blake2;
 extern crate grin_core as core;
@@ -38,5 +39,5 @@ extern crate serde_derive;
 extern crate slog;
 extern crate time;
 
-pub use minimal_pool::MinimalTxPool;
+pub use transaction_pool::TransactionPool;
 pub use types::{BlockChain, PoolAdapter, PoolConfig, PoolError, TxSource};
